+++ test-module/src/test/scala/com/ing/baker/runtime/core/BakerExecutionSpec.scala
@@ -620,7 +620,7 @@ class BakerExecutionSpec extends TestRecipeHelper {
         .withInteractions(interactionOne.withIncrementalBackoffOnFailure(
           initialDelay = 10 milliseconds,
           maximumRetries = 1,
-          exhaustedRetryEvent = Some(exhaustedEvent)))
+          fireExhaustedEvent = true))
 
       when(testInteractionOneMock.apply(anyString(), anyString())).thenThrow(new BakerException())
 
@@ -634,7 +634,7 @@ class BakerExecutionSpec extends TestRecipeHelper {
 
       Thread.sleep(50)
 
-      baker.events(processId).map(_.name) should contain(exhaustedEvent.name)
+      baker.events(processId).map(_.name) should contain(interactionOne.retryExhaustedEventName)
     }
 
     "not fire the exhausted retry event if the interaction passes" in {
@@ -643,7 +643,7 @@ class BakerExecutionSpec extends TestRecipeHelper {
         .withInteractions(interactionOne.withIncrementalBackoffOnFailure(
           initialDelay = 10 milliseconds,
           maximumRetries = 1,
-          exhaustedRetryEvent = Some(exhaustedEvent)))
+          fireExhaustedEvent = true))
 
       val (baker, recipeId) = setupBakerWithRecipe(recipe, mockImplementations)
 
@@ -655,14 +655,14 @@ class BakerExecutionSpec extends TestRecipeHelper {
 
       Thread.sleep(50)
 
-      baker.events(processId).map(_.name) should not contain exhaustedEvent.name
+      baker.events(processId).map(_.name) should not contain interactionOne.retryExhaustedEventName
     }
 
     "fire a specified failure event for an interaction immediately after it fails" in {
 
       val recipe = Recipe("ImmediateFailureEvent")
         .withSensoryEvent(initialEvent)
-        .withInteractions(interactionOne.withFailureStrategy(FireEventAfterFailure(exhaustedEvent)))
+        .withInteractions(interactionOne.withFailureStrategy(FireEventAfterFailure()))
 
       when(testInteractionOneMock.apply(anyString(), anyString())).thenThrow(new RuntimeException("Some failure happened"))
 
@@ -679,9 +679,9 @@ class BakerExecutionSpec extends TestRecipeHelper {
 
       Thread.sleep(50)
       verify(listenerMock).processEvent(processId.toString, RuntimeEvent.create("InitialEvent", Seq("initialIngredient" -> initialIngredientValue)))
-      verify(listenerMock).processEvent(processId.toString, RuntimeEvent.create("RetryExhausted", Seq.empty))
+      verify(listenerMock).processEvent(processId.toString, RuntimeEvent.create(interactionOne.retryExhaustedEventName, Seq.empty))
 
-      baker.events(processId).map(_.name) should contain(exhaustedEvent.name)
+      baker.events(processId).map(_.name) should contain(interactionOne.retryExhaustedEventName)
     }
 
     "be able to return all occurred events" in {